Refactor inventory to items
Renamed the inventory collection to items throughout the codebase for clarity. Updated AdjustStockModal and stocks logic to use isStockIn boolean instead of mode string, and switched quantity parsing from int to float for more accurate stock handling.

+++ src/js/store/collections/stocks.js
@@ -104,8 +104,8 @@ class StocksCollection extends BaseCollection {
   async addStock(itemId, quantity, cost, purchaseDate = new Date()) {
     const stockData = {
       itemId,
-      quantity: parseInt(quantity),
-      remaining: parseInt(quantity),
+      quantity: parseFloat(quantity),
+      remaining: parseFloat(quantity),
       cost: parseFloat(cost),
       purchaseDate: purchaseDate.toISOString(),
       createdAt: new Date().toISOString(),
